Simplify Reasoning action
