+++ front/components/assistant_builder/server_side_props_helpers.ts
@@ -11,6 +11,7 @@ import {
 import { REASONING_MODEL_CONFIGS } from "@app/components/providers/types";
 import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run";
 import type { ProcessConfigurationType } from "@app/lib/actions/process";
+import type { ReasoningConfigurationType } from "@app/lib/actions/reasoning";
 import type { TablesQueryConfigurationType } from "@app/lib/actions/tables_query";
 import type { AgentActionConfigurationType } from "@app/lib/actions/types/agent";
 import {
@@ -22,7 +23,6 @@ import {
   isTablesQueryConfiguration,
   isWebsearchConfiguration,
 } from "@app/lib/actions/types/guards";
-import type { ReasoningConfigurationType } from "@app/lib/actions/types/reasoning";
 import type { RetrievalConfigurationType } from "@app/lib/actions/types/retrieval";
 import { getContentNodesForDataSourceView } from "@app/lib/api/data_source_view";
 import type { Authenticator } from "@app/lib/auth";
