+++ front/lib/actions/runners.ts
@@ -5,6 +5,7 @@ import type { DustAppRunConfigurationType } from "@app/lib/actions/dust_app_run"
 import { DustAppRunConfigurationServerRunner } from "@app/lib/actions/dust_app_run";
 import type { ProcessConfigurationType } from "@app/lib/actions/process";
 import { ProcessConfigurationServerRunner } from "@app/lib/actions/process";
+import type { ReasoningConfigurationType } from "@app/lib/actions/reasoning";
 import { ReasoningConfigurationServerRunner } from "@app/lib/actions/reasoning";
 import { RetrievalConfigurationServerRunner } from "@app/lib/actions/retrieval";
 import { SearchLabelsConfigurationServerRunner } from "@app/lib/actions/search_labels";
@@ -17,7 +18,6 @@ import type {
   BaseActionConfigurationStaticMethods,
 } from "@app/lib/actions/types/base";
 import type { ConversationIncludeFileConfigurationType } from "@app/lib/actions/types/conversation/include_file";
-import type { ReasoningConfigurationType } from "@app/lib/actions/types/reasoning";
 import type { RetrievalConfigurationType } from "@app/lib/actions/types/retrieval";
 import type { SearchLabelsConfigurationType } from "@app/lib/actions/types/search_labels";
 import type { WebsearchConfigurationType } from "@app/lib/actions/types/websearch";
