+++ front/types/assistant/conversation.ts
@@ -1,10 +1,10 @@
 import type { BrowseActionType } from "@app/lib/actions/browse";
 import type { DustAppRunActionType } from "@app/lib/actions/dust_app_run";
 import type { ProcessActionType } from "@app/lib/actions/process";
+import type { ReasoningActionType } from "@app/lib/actions/reasoning";
 import type { TablesQueryActionType } from "@app/lib/actions/tables_query";
 import type { ConversationIncludeFileActionType } from "@app/lib/actions/types/conversation/include_file";
 import type { ConversationListFilesActionType } from "@app/lib/actions/types/conversation/list_files";
-import type { ReasoningActionType } from "@app/lib/actions/types/reasoning";
 import type { RetrievalActionType } from "@app/lib/actions/types/retrieval";
 import type { SearchLabelsActionType } from "@app/lib/actions/types/search_labels";
 import type { WebsearchActionType } from "@app/lib/actions/types/websearch";
